api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
SQLAlchemy==2.0.32
httpx==0.27.0
openai==1.40.2
pandas==2.2.2
numpy==2.0.0
groq==0.32.0